Override HttpCanonicalWebUrlProvider
The `canonicalWebUrl` is used in many places in Gerrit core as well as
in other plugins. It provides the user facing site name that is used to
build various urls.

As in the virtualhost plugin we deal with multiple user facing site
names, we should override the default HTTP `canonicalWebUrl` and return
virtual host name, instead of the host name.

Bug: Issue 299833801
Depends-On: I8321b153782e0bf8131c66ebe477902a434c6dc1
Change-Id: I641a7c7d8b8189cd9f33e188fbdc077bc0282ccb