Honor X-FORWARDED-HOST when negociating server name
In the environemnt where there are multiple hops between Gerrit and a
reverse-proxy server (eg. HTTPS => HAProxy => Gerrit). The HTTP request
server name may not be actually the name of user facing host. For this
purpose a special HTTP header `X-FORWARDED_HOST`[1] is used.

The VirutalHostFilter should first check for that header, then fallback
to the request server name.

[1] https://www.rfc-editor.org/rfc/rfc7239.html#section-5.3

Bug: Issue 300877890
Change-Id: I5a89c216754af7db19b4967f8daaa1c5284b695c