Provide virtual host name in git command outputs
When pushing changes to Gerrit, the URL for each of them will be
returned in the push command output. Previously those URLs used the
name of the main host.

To fix this, our `VirtualHostHttpCanonicalWebUrlProvider` needs to be
aware of the optional request provider, so that we can extract the
server name from the request processed on a different thread where our
`ThreadLocal` was not set. Additionally, it needs to keep its state as
a singleton.

Bug: Issue 311925527
Change-Id: I8259aa99b9924daea9749a999f2fee81c7bf08ed