Do not get HttpServletRequest in non-HTTP threads
When asking for the canonical web url in a non-HTTP thread
it isn't useful to ask for the HttpServletRequest and cause
Guice exceptions: it is way better to just give up and accept
the default behaviour in Gerrit.

Also, when the HttpServletRequest cannot be obtained
the Guice exception to manage is ProvisionException.

Change-Id: Ife5f71285a91c95b161aeb2ce4684ba2261d1aa6