Do not get HttpServletRequest in SSH and SendMail threads
When asking for the canonical web url in a SSH or SendMail
thread it isn't useful to request the HttpServletRequest and
cause Guice exceptions: it is way better to just give up and
accept the default behaviour in Gerrit.

Also, when the HttpServletRequest cannot be obtained
the Guice exception to manage is ProvisionException.

Change-Id: Ife5f71285a91c95b161aeb2ce4684ba2261d1aa6