Implement the PermissionBackend.filterQueryChanges() for queries speedup
When filtering a large site like GerritHub.io and selecting
a small number of projects, it is paramount to reduce the cardinality
of the results returned by Lucene and avoid CPU and memory overload
of post filtering a large number of changes.

Use the projects configuration in virtualhost.config for adding an
extra query filter and slash the number of entries to filter for
permissions.

Change-Id: I2dd78a6053baecd5d37fcce1ae5fe458b23e4a80